requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
sentence-transformers
faiss-cpu
numpy
python-dotenv
google-generativeai
requests